+++ packages/gov-ui-kit/src/components/input/inputTime/index.ts
@@ -1,0 +1,1 @@
+export { InputTime, type IInputTimeProps } from './inputTime';